refactor: restructured file system
