+++ packages/react/src/api/skosmos/SkosApi.ts
@@ -4,15 +4,15 @@ import {
   HierarchyBuilder,
   HierarchyIriProp,
   LoadHierarchyChildrenProps,
-} from "./HierarchyBuilder";
+} from "../../model/interfaces/HierarchyBuilder";
 import {
   Hierarchy,
   ParentChildRelation,
   TreeNode,
-} from "../model/interfaces/Hierarchy";
-import Reified from "../model/Reified";
-import { EntityData } from "../app/types";
-import { dictFromParamString } from "../app/util";
+} from "../../model/interfaces/Hierarchy";
+import Reified from "../../model/Reified";
+import { EntityData } from "../../app";
+import { dictFromParamString } from "../../app/util";
 
 type TopConcept = {
   uri: string;
@@ -169,7 +169,7 @@ export class SkosApi implements HierarchyBuilder {
               entitiesData.set(childNodeData.iri, childNodeData);
             } else {
               if (
-                !childNodeData?.parents?.map((r) => r.value).includes(node.uri)
+                !childNodeData?.parents?.map((r: any) => r.value).includes(node.uri)
               )
                 if (childNodeData?.parents) {
                   childNodeData?.parents?.push(
