+++ packages/react/src/components/widgets/SearchResultsListWidget/MetadataCompact.tsx
@@ -11,7 +11,7 @@ import {
   DescriptionWidget,
   IriWidget,
 } from "../MetadataWidget";
-import { MetadataCompactProps } from "../../../app/types";
+import { MetadataCompactProps } from "../../../app";
 import { pluralizeType } from "../../../app/util";
 import { EntityOntoListPresentation } from "../MetadataWidget/EntityOntoListWidget/EntityOntoListPresentation";
 import { EntityTypeName } from "../../../model/ModelTypeCheck";
@@ -21,7 +21,7 @@ import { Entity } from "../../../model/interfaces";
 import {
   createModelObject,
   getPreferredOntologyJSON,
-} from "../../../model/ModelObjectCreator";
+} from "../../../model/ols-model/ModelObjectCreator";
 import { OlsEntityApi } from "../../../api/ols/OlsEntityApi";
 
 type MetadataInfo = {
