+++ packages/react/src/model/interfaces/Hierarchy.ts
@@ -1,6 +1,6 @@
 import { EntityTypeName } from "../ModelTypeCheck";
-import { HierarchyBuilder } from "../../api/HierarchyBuilder";
-import { EntityData } from "../../app/types";
+import { HierarchyBuilder } from "./HierarchyBuilder";
+import { EntityData } from "../../app";
 
 export type ParentChildRelation = {
   childIri: string;
