+++ packages/react/src/model/ols-model/LinkedEntity.ts
@@ -1,0 +1,9 @@
+export interface LinkedEntity {
+    definedBy?: string[];
+    iri?: string;
+    url?: string;
+    numAppearsIn: string;
+    hasLocalDefinition: boolean;
+    label: string | string[];
+    type: string[];
+}
