+++ packages/react/src/model/ols-model/ModelObjectCreator.ts
@@ -3,16 +3,16 @@ import {
   OLS3Class,
   OLS3Property,
   OLS3Individual,
-} from "./ols3-model";
+} from "../ols3-model";
 import {
   OLS4Ontology,
   OLS4Class,
   OLS4Property,
   OLS4Individual,
-} from "./ols4-model";
-import { Thing } from "./interfaces";
-import { ThingTypeName } from "./ModelTypeCheck";
-import { asArray, inferTypeFromTypeArray } from "../app/util";
+} from "../ols4-model";
+import { Thing } from "../interfaces";
+import { ThingTypeName } from "../ModelTypeCheck";
+import { asArray, inferTypeFromTypeArray } from "../../app/util";
 
 export function createModelObject(response: any) {
   let useLegacy: boolean;
