+++ packages/react/src/model/ols4-model/OLS4Thing.ts
@@ -1,6 +1,6 @@
 import { Thing } from "../interfaces";
 
-import LinkedEntities from "../LinkedEntities";
+import LinkedEntities from "../ols-model/LinkedEntities";
 import Reified from "../Reified";
 
 import { asArray } from "../../app/util";
